save progress

+++ projects/server/src/controller/room.js
@@ -25,14 +25,14 @@ const createRoom = async (req, res) => {
     );
     await transaction.commit();
     res.status(200).send({
-      message: "Success create room type",
+      message: "Success create room.",
       data: result,
     });
   } catch (error) {
     await transaction.rollback();
     console.log("createroomtype", error);
     res.status(500).send({
-      message: "Something wrong on server",
+      message: "Something wrong on server.",
       error,
     });
   }
@@ -48,19 +48,20 @@ const getAllRoom = async (req, res) => {
 
     if (!result) {
       return res.status(200).send({
-        message: "You dont have any room on this property",
+        message: "You dont have any room on this property.",
         data: [],
       });
     }
 
     res.status(200).send({
-      message: "Success get all room from this property",
+      message: "Success get all room from this property.",
       data: result,
     });
   } catch (error) {
     console.log("GETALLROOM", error);
     res.status(500).send({
-      message: "Something wrong on server",
+      message: "Something wrong on server.",
+      error,
     });
   }
 };
@@ -75,18 +76,18 @@ const getOneRoom = async (req, res) => {
 
     if (!result) {
       return res.status(400).send({
-        message: "There is no room found",
+        message: "There is no room found.",
       });
     }
 
     res.status(200).send({
-      message: "Success get room detail",
+      message: "Success get room detail.",
       data: result,
     });
   } catch (error) {
     console.log("getoneroom", error);
     res.status(500).send({
-      message: "Something error on server",
+      message: "Something wrong on server.",
     });
   }
 };
@@ -104,7 +105,7 @@ const updateRoom = async (req, res) => {
 
     if (!room) {
       return res.status(400).send({
-        message: "There is no room found",
+        message: "There is no room found.",
       });
     }
 
@@ -127,24 +128,23 @@ const updateRoom = async (req, res) => {
         base_price: Number(price),
         room_img: imageURL,
       },
-      { where: { id: id, property_id: propId } },
-      { transaction }
+      { where: { id: id, property_id: propId }, transaction }
     );
+    await transaction.commit();
 
     const edited = await db.Room.findOne({
       where: { id: id, property_id: propId },
     });
 
-    await transaction.commit();
     res.status(200).send({
-      message: "Success edit room detail",
+      message: "Success edit room detail.",
       data: edited,
     });
   } catch (error) {
     await transaction.rollback();
     console.log("oneditroom", error);
     res.status(500).send({
-      message: "Something wrong on server",
+      message: "Something wrong on server.",
       error,
     });
   }
@@ -154,12 +154,23 @@ const deleteRoom = async (req, res) => {
   const transaction = await db.sequelize.transaction();
   try {
     const id = req.params.id;
-    const result = await db.Room.destroy(
-      {
-        where: { id: id },
-      },
-      { transaction }
-    );
+    const userId = req.user.id;
+
+    const room = await db.Room.findOne({
+      where: { id: id, deletedAt: null },
+      include: [{ model: db.Property, where: { user_id: userId } }],
+    });
+
+    if (!room) {
+      return res.status(400).send({
+        message: "There is no room found.",
+      });
+    }
+
+    const result = await db.Room.destroy({
+      where: { id: id },
+      transaction,
+    });
 
     await transaction.commit();
     res.status(200).send({
@@ -171,8 +182,8 @@ const deleteRoom = async (req, res) => {
     console.log("ON DELETE ROOM", error);
     res.status(500).send({
       message: "Something wrong on server.",
+      error,
     });
-    error;
   }
 };
